Rename Application to DeckApplication for clarity
