+++ app/src/main/java/it/niedermann/nextcloud/deck/ui/manageaccounts/ManageAccountsActivity.java
@@ -8,12 +8,12 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.nextcloud.android.sso.helper.SingleAccountHelper;
 
-import it.niedermann.nextcloud.deck.Application;
 import it.niedermann.nextcloud.deck.databinding.ActivityManageAccountsBinding;
 import it.niedermann.nextcloud.deck.model.Account;
 import it.niedermann.nextcloud.deck.persistence.sync.SyncManager;
 
-import static it.niedermann.nextcloud.deck.Application.saveCurrentAccountId;
+import static it.niedermann.nextcloud.deck.DeckApplication.readCurrentAccountId;
+import static it.niedermann.nextcloud.deck.DeckApplication.saveCurrentAccountId;
 import static it.niedermann.nextcloud.deck.persistence.sync.adapters.db.util.LiveDataHelper.observeOnce;
 
 public class ManageAccountsActivity extends AppCompatActivity {
@@ -56,7 +56,7 @@ public class ManageAccountsActivity extends AppCompatActivity {
         });
         binding.accounts.setAdapter(adapter);
 
-        observeOnce(syncManager.readAccount(Application.readCurrentAccountId(this)), this, (account -> {
+        observeOnce(syncManager.readAccount(readCurrentAccountId(this)), this, (account -> {
             adapter.setCurrentAccount(account);
             syncManager.readAccounts().observe(this, (localAccounts -> {
                 if (localAccounts.size() == 0) {
