+++ app/src/main/java/it/niedermann/nextcloud/deck/util/ViewUtil.java
@@ -27,10 +27,11 @@ import com.bumptech.glide.request.transition.Transition;
 import java.util.Date;
 import java.util.List;
 
-import it.niedermann.nextcloud.deck.Application;
 import it.niedermann.nextcloud.deck.R;
 import it.niedermann.nextcloud.deck.model.Account;
 import it.niedermann.nextcloud.deck.model.ocs.comment.Mention;
+
+import static it.niedermann.nextcloud.deck.DeckApplication.isDarkTheme;
 
 public final class ViewUtil {
     private ViewUtil() {
@@ -53,7 +54,7 @@ public final class ViewUtil {
         long diff = DateUtil.getDayDifference(new Date(), dueDate);
 
         int backgroundDrawable = 0;
-        int textColor = Application.isDarkTheme(context) ? R.color.dark_fg_primary : R.color.grey600;
+        int textColor = isDarkTheme(context) ? R.color.dark_fg_primary : R.color.grey600;
 
         if (diff == 1) {
             // due date: tomorrow
